pyproject.toml
[project]
name = "ai-quant-trading-platform"
version = "1.0.0"
description = "AI-Driven Quantitative Trading Research Platform"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
dependencies = [
    # Core
    "numpy>=1.26,<3.0",
    "pandas>=2.1",
    "scipy>=1.11",
    "scikit-learn>=1.3",
    # Exchange & Data
    "python-binance>=1.0.19",
    "websockets>=12.0",
    "requests>=2.31",
    # Storage
    "pyarrow>=14.0",
    # Feature Engineering
    "ta>=0.11",
    # Sentiment
    "vaderSentiment>=3.3",
    # Deep Learning
    "torch>=2.1",
    # Reinforcement Learning
    "gymnasium>=0.29",
    "stable-baselines3>=2.2",
    # Visualization
    "streamlit>=1.30",
    "plotly>=5.18",
    # Configuration
    "pydantic>=2.5",
    "pydantic-settings>=2.1",
    "python-dotenv>=1.0",
    # Scheduling
    "apscheduler>=3.10",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "ruff>=0.1",
]

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "W", "F", "I", "B", "UP", "N"]
fixable = ["ALL"]

[tool.ruff.lint.isort]
known-first-party = [
    "config", "utils", "data_pipeline", "feature_engine",
    "feature_store", "market_intelligence", "lstm_models",
    "rl_trader", "alpha_research", "signal_fusion",
    "risk_management", "market_simulator", "execution_engine",
    "orchestrator", "experiments", "scheduler", "dashboard",
]